+++ train.sh
@@ -2,6 +2,9 @@ export CUDA_VISIBLE_DEVICES=
 
 
 python -m bin.train \
+  --problem_name "heheda" \
+  --train "
+    pretrain_model: pretrained_model" \
   --config_paths "
         ./njunmt/example_configs/toy_seq2seq.yml,
         ./default_configs/default_optimizer.yml,