+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - WMT2017 EN<->DE benchmark.
+- Load pretrained model.
 
 ### Changed
 - Default loss function.